[change] uint64 => int
落ち着いて考えてみるとそもそもDB側がINT型なのでint32以外を指定するべきではなかった。（扱いやすいのでintを採用した）

+++ src/repository/repository.go
@@ -13,9 +13,9 @@ type repository struct {
 
 type Repository interface {
 	CreateNewTask(title string, content string) (*models.Task, error)
-	FindTaskByID(id uint64) (*models.Task, error)
-	UpdateTask(id uint64, title string, content string) error
-	DeleteTask(id uint64) error
+	FindTaskByID(id int) (*models.Task, error)
+	UpdateTask(id int, title string, content string) error
+	DeleteTask(id int) error
 	Close() error
 }
 
@@ -33,8 +33,8 @@ func (repo *repository) Close() error {
 	return repo.db.Close()
 }
 
-func (repo *repository) generateID() (uint64, error) {
-	var id uint64
+func (repo *repository) generateID() (int, error) {
+	var id int
 	err := repo.db.Get(
 		&id,
 		`
